requirements.txt
supabase==2.18.0
python-dotenv==1.1.1
requests==2.32.4
pydantic==2.11.7